// apps/nk-vrapce/src/types/hns.ts
export interface MatchPhase {
  phaseTypeId: number | null;
  name: string | null;
  fcdName: string | null;
}

export interface HnsFacility {
  id: number | null;
  fifaId: string | null;
  name: string | null;
  address: string | null;
  place: string | null;
  latitude: number | null;
  longitude: number | null;
}

export interface HnsEventType {
  eventTypeId: number | null;
  name: string | null;
  fcdName: string | null;
}

export interface HnsEventPlayer {
  roleId: number | null;
  personId: number | null;
  fifaId: string | null;
  name: string | null;
  shortName: string | null;
  shirtNumber: number | null;
  picture: string | null;
  hideProfile: boolean | null;
}
// ...
export interface HnsMatchEvent {
  eventId: number | null;
  eventType: HnsEventType | null;
  matchPhase: MatchPhase | null;
  minute: number | null;
  minuteFull: number | null;
  stoppageTime: number | null;
  displayMinute: string | null;
  player: HnsEventPlayer | null;
  player2: HnsEventPlayer | null;
  homeTeam: boolean | null;
  orderNumber: number | null;
}